requirements.txt
astroid==2.2.5
colorama==0.4.1
Cython==0.29.6
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==6.0.0
pylint==2.3.1
pytesseract==0.2.6
selenium==3.141.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
